Enable table ojbects to be called as a function
When a table object is called as a function, `get` method is called.
The method hehaves as a function. It describes a relation ship
between input and output and it has no side effect.

I intend `get` method is used mainly, so I think it can be written
with a brief notation.

This feature, however, might be removed in the future if it seems to
be unnatural.