Enable to use Table at for-loops
A table object can behave as a container or an iterator. It returns
all applicable rows.

This feature is intended to be used, for example, in testing.
Parameters of test cases are described in a table, and tests ara run
in a for-loop. Of cource the same logic can be witten with a list of
lists (i.e., a 2D allay), but a table formated pretty is better,
isn't it?

Additionally, i-th row can get with table[index] (__getitem__).
It raises LookupError if the row has N/A or Wild Card. That is, the
table rases an error at an in-bound index. If this behavior seems to
be unnatural, this specification might be changed.