Support reStructuredText Grid Table format
The table format is estimated automatically. Manual setting should be
avoided because this module is for human. The goal is programming free
programming.

Change the table text pareser to plugable. When we want to add a format,
add a new member to Format class.

Format.REST_GRID_TABLE is added for the support of GridTable . Because
REST_SIMPLE_TABLE and REST_GRID_TABLE has common logic, the common logic
is written in Format._ReSTTable class.