Save row values in Table class as namedtuples
Than save as list the code is smarter if we use namedtuples. We do not
need search the column number from a lebel name.

Moreover Table.label field is to be unnecessary. Instead of the field,
I provide Table._label() method. This method is attached @property
decorator. We can get labels with table.label, do not need ().

I attached @property also to Table._num_columns() and Table.num_rows().