Rename "Attribute" class to "ColumnType"
The name of "Attrubute" is confusing. Attrubute also means a name of
a object member or a tuple member (We call it label in this module).

"Property" is considered. It means, however, a value of attrubute
(especially in Python?).

"ColumnType" is good naming, I think. The values in column type are
used as type objects. I found this when I got the name. I learned the
fact naming refines the design.